pyproject.toml
[tool.poetry]
name = "cloudpayments"
version = "0.1.0"
description = "Test task"
authors = ["Aleksandr Osovskii"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
marshmallow-dataclass = "^8.5.8"
black = "^22.6.0"
aiohttp-basicauth = "^1.0.0"
python-environ = "^0.4.54"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
